fix : 저번 PR에 대한 Interface 부분을 fix 원상 복구했습니다. 또한 CorporateEntity와 UserEntity의 package를 합쳤습니다. 둘다 회원이지만 다른 회원이기 때문에, 공통으로 사용하는 서비스가 많아 같은 package에서 UserService Interface를 Overriding 하여 사용할 것입니다.
