+++ src/main/java/kr/co/popoolserver/common/infra/jwt/JwtProvider.java
@@ -7,13 +7,12 @@ import kr.co.popoolserver.common.infra.error.exception.JwtTokenInvalidException;
 import kr.co.popoolserver.common.infra.error.exception.NotFoundException;
 import kr.co.popoolserver.common.infra.error.exception.UserDefineException;
 import kr.co.popoolserver.common.infra.error.model.ErrorCode;
-import kr.co.popoolserver.user.domain.UserEntity;
+import kr.co.popoolserver.user.domain.entity.UserEntity;
 import kr.co.popoolserver.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
