--- src/main/java/kr/co/popoolserver/corporate/controller/CorporateController.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.corporate.controller;
-
-public class CorporateController {
-}
