--- src/main/java/kr/co/popoolserver/corporate/domain/CorporateDto.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.corporate.domain;
-
-public class CorporateDto {
-}
