--- src/main/java/kr/co/popoolserver/corporate/domain/service/CorporateService.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.corporate.domain.service;
-
-public interface CorporateService {
-}
