--- src/main/java/kr/co/popoolserver/corporate/domain/service/CorporateServiceImpl.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.corporate.domain.service;
-
-public class CorporateServiceImpl {
-}
