--- src/main/java/kr/co/popoolserver/corporate/repository/CorporateRepository.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.corporate.repository;
-
-public interface CorporateRepository {
-}
