+++ src/main/java/kr/co/popoolserver/user/controller/CorporateController.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.user.controller;
+
+public class CorporateController {
+}
