+++ src/main/java/kr/co/popoolserver/user/domain/dto/corporateDto/CorporateDto.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.user.domain.dto.corporateDto;
+
+public class CorporateDto {
+}
