+++ src/main/java/kr/co/popoolserver/user/domain/dto/userDto/UserCreateDto.java
@@ -1,4 +1,4 @@
-package kr.co.popoolserver.user.domain.dto;
+package kr.co.popoolserver.user.domain.dto.userDto;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
