+++ src/main/java/kr/co/popoolserver/user/domain/service/UserService.java
@@ -1,28 +1,28 @@
 package kr.co.popoolserver.user.domain.service;
 
-import kr.co.popoolserver.user.domain.dto.UserCreateDto;
+import kr.co.popoolserver.user.domain.dto.userDto.UserCreateDto;
 
 public interface UserService {
 
     //login
-    default UserCreateDto.TOKEN login(UserCreateDto.LOGIN login){ return null; }
+    UserCreateDto.TOKEN login(UserCreateDto.LOGIN login);
 
     //create
-    default void signUp(UserCreateDto.CREATE create){}
+    void signUp(UserCreateDto.CREATE create);
 
     //update
-    default void updateUser(UserCreateDto userCreateDto){}
+    void updateUser(UserCreateDto userCreateDto);
 
     //get
-    default void getUser(){}
+    void getUser();
 
     //delete
-    default void deleteUser(){}
+    void deleteUser();
 
     //common
-    default void checkIdentity(String identity){}
-    default void checkPhoneNumber(String phoneNumber){}
-    default void checkPassword(String password, String checkPassword){}
-    default void checkEncodePassword(String password, String encodePassword){}
-    default void checkDelete(String delYN){}
+    void checkIdentity(String identity);
+    void checkPhoneNumber(String phoneNumber);
+    void checkPassword(String password, String checkPassword);
+    void checkEncodePassword(String password, String encodePassword);
+    void checkDelete(String delYN);
 }
