+++ src/main/java/kr/co/popoolserver/user/domain/service/impl/CorporateServiceImpl.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.user.domain.service.impl;
+
+public class CorporateServiceImpl {
+}
