--- src/main/java/kr/co/popoolserver/user/domain/service/impl/UserCommonServiceImpl.java
@@ -1,66 +1,0 @@
-package kr.co.popoolserver.user.domain.service.impl;
-
-import kr.co.popoolserver.common.domain.PhoneNumber;
-import kr.co.popoolserver.common.infra.error.exception.BusinessLogicException;
-import kr.co.popoolserver.common.infra.error.exception.DuplicatedException;
-import kr.co.popoolserver.common.infra.error.model.ErrorCode;
-import kr.co.popoolserver.user.domain.service.UserService;
-import kr.co.popoolserver.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserCommonServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    /**
-     * ID duplicated check
-     * @param identity
-     */
-    @Override
-    public void checkIdentity(String identity) {
-        if(userRepository.existsByIdentity(identity)) throw new DuplicatedException(ErrorCode.DUPLICATED_ID);
-    }
-
-    /**
-     * Phone duplicated check
-     * @param phoneNumber
-     */
-    @Override
-    public void checkPhoneNumber(String phoneNumber) {
-        if(userRepository.existsByPhoneNumber(new PhoneNumber(phoneNumber))) throw new DuplicatedException(ErrorCode.DUPLICATED_PHONE);
-    }
-
-    /**
-     * PW check
-     * @param password : use pw
-     * @param checkPassword : check pw
-     */
-    @Override
-    public void checkPassword(String password, String checkPassword) {
-        if(!password.equals(checkPassword)) throw new BusinessLogicException(ErrorCode.WRONG_PASSWORD);
-    }
-
-    /**
-     * Login PW Check
-     * @param password
-     * @param encodePassword
-     */
-    @Override
-    public void checkEncodePassword(String password, String encodePassword) {
-        if(!passwordEncoder.matches(password, encodePassword)) throw new BusinessLogicException(ErrorCode.WRONG_PASSWORD);
-    }
-
-    /**
-     * delete Check
-     * @param delYN
-     */
-    @Override
-    public void checkDelete(String delYN) {
-        if(delYN.equals("Y")) throw new BusinessLogicException(ErrorCode.DELETED_USER);
-    }
-}
