--- src/main/java/kr/co/popoolserver/user/domain/service/impl/UserDeleteServiceImpl.java
@@ -1,18 +1,0 @@
-package kr.co.popoolserver.user.domain.service.impl;
-
-import kr.co.popoolserver.user.domain.service.UserService;
-import kr.co.popoolserver.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserDeleteServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public void deleteUser() {
-        //TODO: delete user
-    }
-}
