+++ src/main/java/kr/co/popoolserver/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package kr.co.popoolserver.user.repository;
 
 import kr.co.popoolserver.common.domain.PhoneNumber;
-import kr.co.popoolserver.user.domain.UserEntity;
+import kr.co.popoolserver.user.domain.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
